// nextjs/gbiz-next/src/packages/verification/src/api/risk-validation-center/types.gen.ts
export type GenericResult_RatingPriorityResponse_ = {
    code?: string;
    data?: RatingPriorityResponse;
    msg?: string;
    retry?: boolean;
    success?: boolean;
};
// ...
export type GetValidationsUsingPostResponses = {
    /**
     * OK
     */
    200: GenericResult_RatingPriorityResponse_;
};

export type GetValidationsUsingPostResponse = GetValidationsUsingPostResponses[keyof GetValidationsUsingPostResponses];